GafferSceneModule : Split IECoreScenePreviewBinding from RenderBinding
